+++ familyhelper-ui-view/packages/webapp-mobile/src/navigation/modules/welcome.ts
@@ -37,7 +37,7 @@ const navigationNodes: NavigationNodeSetting[] = [
     router: {
       required: true,
       path: 'welcome',
-      component: () => import('@/views/nodes/welcome/Welcome.vue'),
+      component: () => import('@/views/nodes/simple/welcome/Welcome.vue'),
     },
     permission: {
       required: false,